# Conflicts:
#	src/main/java/com/example/tarea4_grupo2/dto/PedidosReporteDto.java

+++ src/main/java/com/example/tarea4_grupo2/dto/PedidosDisponiblesDTO.java
@@ -6,6 +6,6 @@ public interface PedidosDisponiblesDTO {
     String getDistritorestaurante();
     String getDireccioncliente();
     int getComision();
-    int getMonto();
+    float getMonto();
 
 }
